implement iface speed detection on Linux (closes #12)
This follows the deprecated SIOCETHTOOL interface, but should be
generally more portable. Was tested on Debian GNU/Linux "buster" with
a 4.18.0 Linux kernel.

Note that the interface is similar to the the BSD media.h pattern
(create a socket and pass it to ioctl) but I felt it was clearer to
just create another #if block than mangle the already quite complex

This was mostly cargo-culted from:

https://stackoverflow.com/questions/2872058/get-link-speed-programmatically